+++ packages/desktop/src/rtl-tests/community.create.test.tsx
@@ -73,7 +73,7 @@ describe('User', () => {
       if (action === SocketActionTypes.CREATE_NETWORK) {
         const data = input[1] as Community
         const payload = { ...data, privateKey: 'privateKey' }
-        socket.socketClient.emit<ResponseCreateNetworkPayload>(SocketActionTypes.NETWORK, {
+        socket.socketClient.emit<ResponseCreateNetworkPayload>(SocketActionTypes.NETWORK_CREATED, {
           community: payload,
           network: {
             hiddenService: {
@@ -88,7 +88,7 @@ describe('User', () => {
       }
       if (action === SocketActionTypes.REGISTER_OWNER_CERTIFICATE) {
         const payload = input[1] as RegisterOwnerCertificatePayload
-        socket.socketClient.emit<SavedOwnerCertificatePayload>(SocketActionTypes.SAVED_OWNER_CERTIFICATE, {
+        socket.socketClient.emit<SavedOwnerCertificatePayload>(SocketActionTypes.OWNER_CERTIFICATE_ISSUED, {
           communityId: payload.communityId,
           network: {
             certificate: payload.permsData.certificate,
@@ -97,14 +97,14 @@ describe('User', () => {
       }
       if (action === SocketActionTypes.CREATE_COMMUNITY) {
         const payload = input[1] as InitCommunityPayload
-        socket.socketClient.emit<ResponseLaunchCommunityPayload>(SocketActionTypes.COMMUNITY, {
+        socket.socketClient.emit<ResponseLaunchCommunityPayload>(SocketActionTypes.COMMUNITY_LAUNCHED, {
           id: payload.id,
         })
-        socket.socketClient.emit(SocketActionTypes.NEW_COMMUNITY, {
+        socket.socketClient.emit(SocketActionTypes.COMMUNITY_CREATED, {
           id: payload.id,
         })
 
-        socket.socketClient.emit<ChannelsReplicatedPayload>(SocketActionTypes.CHANNELS_REPLICATED, {
+        socket.socketClient.emit<ChannelsReplicatedPayload>(SocketActionTypes.CHANNELS_STORED, {
           channels: {
             general: {
               name: 'general',