+++ packages/state-manager/src/types.ts
@@ -49,10 +49,10 @@ export interface EmitEvents {
   [SocketActionTypes.CLOSE]: () => void
   [SocketActionTypes.LEAVE_COMMUNITY]: () => void
   [SocketActionTypes.CREATE_NETWORK]: EmitEvent<Community>
-  [SocketActionTypes.SAVE_USER_CSR]: EmitEvent<SaveCSRPayload>
-  [SocketActionTypes.SEND_COMMUNITY_METADATA]: EmitEvent<CommunityMetadata, (response: CommunityMetadata) => void>
-  [SocketActionTypes.SEND_COMMUNITY_CA_DATA]: EmitEvent<PermsData>
-  [SocketActionTypes.SAVE_USER_PROFILE]: EmitEvent<UserProfile>
+  [SocketActionTypes.ADD_CSR]: EmitEvent<SaveCSRPayload>
+  [SocketActionTypes.SET_COMMUNITY_METADATA]: EmitEvent<CommunityMetadata, (response: CommunityMetadata) => void>
+  [SocketActionTypes.SET_COMMUNITY_CA_DATA]: EmitEvent<PermsData>
+  [SocketActionTypes.SET_USER_PROFILE]: EmitEvent<UserProfile>
 }
 
 export type Socket = IOSocket<EventsMap, EmitEvents>