+++ packages/types/src/socket.ts
@@ -1,64 +1,80 @@
+/**
+ * Backend API event types. Currently, these are divided into two
+ * groups: pure events and actions. Pure events are emitted from the
+ * backend to notify the frontend of something and are generally named
+ * with the past tense (e.g. COMMUNITY_CREATED), while actions are
+ * emitted from the frontend in order to invoke the backend to do
+ * something on it's behalf and are generally named as a command (e.g.
+ * CREATE_COMMUNITY).
+ */
 export enum SocketActionTypes {
-  // C
-  CANCEL_DOWNLOAD = 'cancelDownload',
-  CHANNELS_REPLICATED = 'channelsReplicated',
-  CHANNEL_SUBSCRIBED = 'channelSubscribed',
-  CLOSE = 'close',
-  COMMUNITY = 'community',
-  COMMUNITY_METADATA_LOADED = 'communityMetadataLoaded',
-  CONNECTED_PEERS = 'connectedPeers',
-  CONNECT_TO_WEBSOCKET_SERVER = 'connectToWebsocketServer',
-  CONNECTION = 'connection',
-  CREATE_CHANNEL = 'createChannel',
+  // ====== Community ======
+
+  COMMUNITY_CREATED = 'communityCreated',
+  COMMUNITY_LAUNCHED = 'communityLaunched',
+  COMMUNITY_METADATA_STORED = 'communityMetadataStored',
   CREATE_COMMUNITY = 'createCommunity',
-  CREATE_NETWORK = 'createNetwork',
-  // D
-  DIRECT_MESSAGE = 'directMessage',
-  DOWNLOAD_FILE = 'downloadFile',
-  DOWNLOAD_PROGRESS = 'downloadProgress',
-  DELETE_CHANNEL = 'deleteChannel',
-  DELETE_FILES_FROM_CHANNEL = 'deleteFilesFromChannel',
-  // E
-  ERROR = 'error',
-  // G
-  GET_MESSAGES = 'getMessages',
-  // L
   LAUNCH_COMMUNITY = 'launchCommunity',
   LEAVE_COMMUNITY = 'leaveCommunity',
-  LIBP2P_PSK_SAVED = 'libp2pPskSaved',
-  LOADED_USER_PROFILES = 'loadedUserProfiles',
-  // M
-  MESSAGE_IDS_LOADED = 'messageIdsLoaded',
-  MESSAGES_LOADED = 'messagesLoaded',
-  // N
-  NETWORK = 'network',
-  NEW_COMMUNITY = 'newCommunity',
-  // P
+  SET_COMMUNITY_CA_DATA = 'setCommunityCaData',
+  SET_COMMUNITY_METADATA = 'setCommunityMetadata',
+
+  // ====== Channels ======
+
+  CHANNEL_SUBSCRIBED = 'channelSubscribed',
+  CHANNELS_STORED = 'channelsStored',
+  CREATE_CHANNEL = 'createChannel',
+  DELETE_CHANNEL = 'deleteChannel',
+  DELETE_FILES_FROM_CHANNEL = 'deleteFilesFromChannel',
+
+  // ====== Messages ======
+
+  GET_MESSAGES = 'getMessages',
+  MESSAGE_IDS_STORED = 'messageIdsStored',
+  MESSAGE_MEDIA_UPDATED = 'messageMediaUpdated',
+  MESSAGES_STORED = 'messagesStored',
+  SEND_MESSAGE = 'sendMessage',
+
+  // ====== User ======
+
+  SET_USER_PROFILE = 'updateUserProfile',
+  USER_PROFILES_STORED = 'userProfilesStored',
+
+  // ====== Files ======
+
+  CANCEL_DOWNLOAD = 'cancelDownload',
+  CHECK_FOR_MISSING_FILES = 'checkForMissingFiles',
+  DOWNLOAD_FILE = 'downloadFile',
+  DOWNLOAD_PROGRESS = 'downloadProgress',
+  FILE_UPLOADED = 'fileUploaded',
+  REMOVE_DOWNLOAD_STATUS = 'removeDownloadStatus',
+  UPLOAD_FILE = 'uploadFile',
+
+  // ====== Certificates ======
+
+  ADD_CSR = 'addCsr',
+  CERTIFICATES_STORED = 'certificatesStored',
+  CSRS_STORED = 'csrsStored',
+  OWNER_CERTIFICATE_ISSUED = 'ownerCertificateIssued',
+  REGISTER_USER_CERTIFICATE = 'registerUserCertificate',
+  REGISTER_OWNER_CERTIFICATE = 'registerOwnerCertificate',
+
+  // ====== Network ======
+
+  CLOSE = 'close',
+  CONNECTED_PEERS = 'connectedPeers',
+  CONNECTION = 'connection',
+  CONNECTION_PROCESS_INFO = 'connectionProcess',
+  CREATE_NETWORK = 'createNetwork',
+  LIBP2P_PSK_STORED = 'libp2pPskStored',
+  NETWORK_CREATED = 'networkCreated',
   PEER_CONNECTED = 'peerConnected',
   PEER_DISCONNECTED = 'peerDisconnected',
   PEER_LIST = 'peerList',
-  PUSH_NOTIFICATION = 'pushNotification',
-  // R
-  REGISTER_USER_CERTIFICATE = 'registerUserCertificate',
-  REGISTER_OWNER_CERTIFICATE = 'registerOwnerCertificate',
-  REMOVE_DOWNLOAD_STATUS = 'removeDownloadStatus',
-  RESPONSE_GET_CERTIFICATES = 'responseGetCertificates',
-  RESPONSE_GET_CSRS = 'responseGetCsrs',
-  REQUEST_PEER_ID = 'requestPeerId',
-  // S
-  SAVED_OWNER_CERTIFICATE = 'savedOwnerCertificate',
-  SAVE_USER_CSR = 'saveUserCsr',
-  SAVE_USER_PROFILE = 'saveUserProfile',
-  SAVED_USER_CSR = 'savedUserCsr',
-  SEND_MESSAGE = 'sendMessage',
-  SEND_PEER_ID = 'sendPeerId',
-  SEND_COMMUNITY_METADATA = 'sendCommunityMetadata',
-  SEND_COMMUNITY_CA_DATA = 'sendCommunityCaData',
-  //  T
   TOR_INITIALIZED = 'torInitialized',
-  CONNECTION_PROCESS_INFO = 'connectionProcess',
-  // U
-  UPDATE_MESSAGE_MEDIA = 'updateMessageMedia',
-  UPLOAD_FILE = 'uploadFile',
-  UPLOADED_FILE = 'uploadedFile',
+
+  // ====== Misc ======
+
+  PUSH_NOTIFICATION = 'pushNotification',
+  ERROR = 'error',
 }